10th commit (created models.py, run.py and views.py. Modified a bit text formulation in bill.html)
